fix: harden HITL review fingerprint resume paths
Allow forgejo-reviewer-summary live refresh, fail closed on provider
errors (not-found only marks missing), include full non-Looper thread
fingerprints so mid-park replies invalidate parked answers, document
fingerprint trade-offs, and split oversized HITL contract tests.

Generated-By: looper 0.11.2 (runner=fixer, agent=grok-build)
